Refactor the call stack state functions.
