+++ src/stackmap_checker/jump.s
@@ -1,7 +1,7 @@
-.global restore_and_jmp
+.global jmp_to_addr
 
-# Restores the register state stored in array r and jumps to addr
-restore_and_jmp:
+# Jumps to addr, cleaning up after __guard_failure
+jmp_to_addr:
     mov    %rbp, %rsp
     pop    %rbp
     pop    %rax # pop the return address of __guard_failure
